+++ packages/contact-center/cc-components/src/components/task/TaskList/task-list.tsx
@@ -4,12 +4,13 @@ import Task from '../Task';
 import './styles.scss';
 
 const TaskListComponent: React.FunctionComponent<TaskListComponentProps> = (props) => {
-  const {currentTask, taskList, acceptTask, declineTask, isBrowser, onTaskSelect} = props;
+  const {currentTask, taskList, acceptTask, declineTask, isBrowser, onTaskSelect, logger} = props;
+
   if (!taskList || Object.keys(taskList).length === 0) {
     return <></>; // hidden component
   }
   return (
-    <ul className="task-list">
+    <ul className="task-list" data-testid="task-list">
       {Object.values(taskList)?.map((task, index) => {
         const callAssociationDetails = task?.data?.interaction?.callAssociatedDetails;
         const ani = callAssociationDetails?.ani;
@@ -32,6 +33,10 @@ const TaskListComponent: React.FunctionComponent<TaskListComponentProps> = (prop
             : undefined;
         const declineText =
           isIncomingTask && !task.data.wrapUpRequired && isTelephony && isBrowser ? 'Decline' : undefined;
+        logger.info('CC-Widgets: TaskList: rendering task list', {
+          module: 'task-list.tsx',
+          method: 'renderItem',
+        });
         return (
           <Task
             interactionId={task.data.interactionId}
@@ -46,6 +51,10 @@ const TaskListComponent: React.FunctionComponent<TaskListComponentProps> = (prop
             declineTask={() => declineTask(task)}
             ronaTimeout={isIncomingTask ? ronaTimeout : null}
             onTaskSelect={() => {
+              logger.log(`CC-Widgets: TaskList: select task clicked for interactionId: ${task.data.interactionId}`, {
+                module: 'task-list.tsx',
+                method: 'onTaskSelect',
+              });
               if (
                 currentTask?.data.interactionId !== task.data.interactionId &&
                 !(isIncomingTask && !task.data.wrapUpRequired)