+++ packages/contact-center/cc-components/tests/components/task/CallControl/consult-transfer-popover.tsx
@@ -4,6 +4,14 @@ import {render, screen, fireEvent} from '@testing-library/react';
 import '@testing-library/jest-dom';
 import ConsultTransferPopoverComponent from '../../../../src/components/task/CallControl/CallControlCustom/consult-transfer-popover';
 
+const loggerMock = {
+  log: jest.fn(),
+  info: jest.fn(),
+  warn: jest.fn(),
+  trace: jest.fn(),
+  error: jest.fn(),
+};
+
 jest.mock('../../../../src/components/task/CallControl/CallControlCustom/consult-transfer-list-item', () => {
   const MockListItem = (props: any) => (
     <div data-testid="ConsultTransferListComponent" onClick={props.onButtonPress}>
@@ -12,36 +20,6 @@ jest.mock('../../../../src/components/task/CallControl/CallControlCustom/consult
   );
   MockListItem.displayName = 'ConsultTransferListComponent';
   return MockListItem;
-});
-
-jest.mock('@momentum-design/components/dist/react', () => ({
-  Icon: (props: any) => <div {...props} />,
-}));
-
-jest.mock('@momentum-ui/react-collaboration', () => {
-  return {
-    Text: (props: any) => <div data-testid="Text" {...props} />,
-    TabListNext: (props: any) => {
-      if (props['tab-select-trigger'] && props.onTabSelection) {
-        setTimeout(() => props.onTabSelection(props['tab-select-trigger']), 0);
-      }
-      return (
-        <div data-testid="TabListNext" {...props}>
-          {props.children}
-        </div>
-      );
-    },
-    TabNext: (props: any) => (
-      <div data-testid="TabNext" {...props}>
-        {props.children}
-      </div>
-    ),
-    ListNext: (props: any) => (
-      <div data-testid="ListNext" {...props}>
-        {props.children}
-      </div>
-    ),
-  };
 });
 
 beforeAll(() => {
@@ -68,6 +46,7 @@ describe('ConsultTransferPopoverComponent', () => {
     ],
     onAgentSelect: mockOnAgentSelect,
     onQueueSelect: mockOnQueueSelect,
+    logger: loggerMock,
   };
 
   beforeEach(() => {