+++ packages/contact-center/task/tests/CallControl/index.tsx
@@ -2,13 +2,8 @@ import React from 'react';
 import {render} from '@testing-library/react';
 import * as helper from '../../src/helper';
 import {CallControl} from '../../src';
+import store from '@webex/cc-store';
 import '@testing-library/jest-dom';
-
-jest.mock('@webex/cc-components', () => {
-  return {
-    CallControlComponent: () => <div>CallControlComponent</div>,
-  };
-});
 
 // Mock the store
 jest.mock('@webex/cc-store', () => ({
@@ -16,7 +11,12 @@ jest.mock('@webex/cc-store', () => ({
   deviceType: 'BROWSER',
   dialNumber: '12345',
   wrapupCodes: [],
-  logger: {},
+  logger: {
+    log: jest.fn(),
+    error: jest.fn(),
+    warn: jest.fn(),
+    info: jest.fn(),
+  },
   currentTask: {
     data: {interactionId: 'mockInteractionId'},
     on: jest.fn(),
@@ -42,7 +42,7 @@ describe('CallControl Component', () => {
   it('renders CallControlPresentational with correct props', () => {
     const useCallControlSpy = jest.spyOn(helper, 'useCallControl').mockReturnValue({
       currentTask: {
-        data: {interactionId: 'mockInteractionId'},
+        data: {interactionId: 'mockInteractionId', interaction: {mediaType: 'voice'}},
         on: jest.fn(),
         off: jest.fn(),
         hold: jest.fn(),
@@ -69,6 +69,20 @@ describe('CallControl Component', () => {
       consultInitiated: undefined,
       featureFlags: undefined,
       deviceType: 'BROWSER',
+      controlVisibility: {
+        accept: false,
+        decline: false,
+        end: false,
+        muteUnmute: false,
+        holdResume: true,
+        consult: false,
+        transfer: false,
+        conference: false,
+        wrapup: false,
+        pauseResumeRecording: false,
+        endConsult: false,
+        recordingIndicator: false,
+      },
     });
 
     render(<CallControl onHoldResume={onHoldResumeCb} onEnd={onEndCb} onWrapUp={onWrapUpCb} />);
@@ -89,7 +103,7 @@ describe('CallControl Component', () => {
       onHoldResume: expect.any(Function),
       onEnd: expect.any(Function),
       onWrapUp: expect.any(Function),
-      logger: {},
+      logger: store.logger,
       consultInitiated: undefined,
       featureFlags: undefined,
       deviceType: 'BROWSER',