+++ packages/contact-center/user-state/bable.config.js
@@ -1,0 +1,3 @@
+const baseConfig = require('../../../babel.config.js');
+
+module.exports = baseConfig;