+++ packages/contact-center/user-state/tests/user-state/index.tsx
@@ -3,11 +3,6 @@ import {render} from '@testing-library/react';
 import {UserState} from '../../src';
 import * as helper from '../../src/helper';
 import '@testing-library/jest-dom';
-jest.mock('@webex/cc-components', () => {
-  return {
-    UserStateComponent: () => <div>UserStateComponent</div>,
-  };
-});
 
 // Mock the store import
 jest.mock('@webex/cc-store', () => {
@@ -20,6 +15,7 @@ jest.mock('@webex/cc-store', () => {
     agentId: 'testAgentId',
     logger: {
       log: jest.fn(),
+      info: jest.fn(),
     },
     lastStateChangeTimestamp: new Date(),
     customState: null,
@@ -62,6 +58,7 @@ describe('UserState Component', () => {
       lastStateChangeTimestamp: expect.any(Date),
       logger: {
         log: expect.any(Function),
+        info: expect.any(Function),
       },
       onStateChange: expect.any(Function),
     });